bug fixed, removed useless iss site
